revert tooltip refactor
this reverts commit d0004a182

now we do not render the TooltipTrigger if it's disabled so we can revert back
to the better onOpenChange callback to intercept and check if content is
truncated or not

otherwise it can interfere with other onPointerMove event handlers on the
element
