+++ packages/react/src/tooltip/TooltipContext.ts
@@ -1,9 +1,11 @@
 "use client";
 
+import type { RefObject } from "react";
+
 import { createContext } from "@radix-ui/react-context";
 
 export const [TooltipProvider, useTooltipContext] = createContext<{
-  auto: boolean | undefined;
   open: boolean | undefined;
   setOpen: (open: boolean) => void;
+  triggerRef: RefObject<HTMLButtonElement>;
 }>("@optiaxiom/react/Tooltip");
